refactor: more iterator trickery
We can avoid manually tracking the buffer if we just wrap the *reader*
in an Arc<Mutex<>>, rather than wrapping the iterator in it. This also
means that every time we change tracks we have to start a new cpal
stream, which is annoying, but we can figure that out later.